% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsely.R
\name{parse_genes}
\alias{parse_genes}
\title{Get list of genes}
\usage{
parse_genes(x, keep = NULL, sep = ",")
}
\arguments{
\item{x}{character containing genes to grab}

\item{keep}{character list of genes to keep}

\item{sep}{seperator of gene names}
}
\value{
List of genes. Only those in keep if keep is not NULL
}
\description{
Takes a single string of gene names and turns them into a list
}
\examples{
parse_genes("Xist,cdk9,Est3, Pea3")
}
