% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsely.R
\name{parse_gsea_tsv}
\alias{parse_gsea_tsv}
\title{Gets list of genes from gsea tsv}
\usage{
parse_gsea_tsv(x)
}
\arguments{
\item{x}{String path to tsv}
}
\value{
List of genes from tsv
}
\description{
Gets list of genes from gsea tsv
}
\examples{
#parse_gsea_tsv("../data/BIOCARTA_ETS_PATHWAY.v2023.2.Mm.tsv") Get more complete examples
}
