% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsely.R
\name{parse_url_raw}
\alias{parse_url_raw}
\title{Grab all genes from a url text file}
\usage{
parse_url_raw(x, re = "(?<=(\\\\.symbol\\\\.))[:alnum:]+")
}
\arguments{
\item{x}{String url}

\item{re}{Regex expression to use, default looks for the text ".symbol." preceding the gene symbol}
}
\value{
A character vector of all found genes
}
\description{
Grab all genes from a url text file
}
\examples{
#Fill later
}
